requirements.txt
python-telegram-bot
pandas
apscheduler
tenacity
python-dotenv
pytz
httpx
Pillow
aiofiles
colorama
httpx[http2]
